# src/swagger-config.json
{
    "swagger": "2.0",
    "info": {
      "description": "API do sistema de Licenciamento de Ambulantes de Salvador",
      "version": "1.1.0",
      "title": "LAS API",
      "contact": {
        "email": "admin@las.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "virtserver.swaggerhub.com",
    "basePath": "/tiagolpadua/las-api/1.1.0",
...
    "schemes": [
      "https"
    ],
    "paths": {
      "/usuarios": {
        "get": {
          "tags": [
            "usuario"
          ],
          "summary": "Lista os usuarios",
          "description": "Lista os usuários do LAS-API",
          "operationId": "listarUsuarios",
          "produces": [
            "application/json"
          ],
          "parameters": [],
          "responses": {
            "200": {
              "description": "Resultado da listagem",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Usuario"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "usuario"
          ],
          "summary": "Inclui um usuário",
          "description": "Inclui um usuário no LAS-API",
          "operationId": "incluirUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "usuario",
              "description": "Usuário a ser incluído",
              "required": false,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Usuário incluído com sucesso"
            },
            "400": {
              "description": "Entrada inválida"
            }
          }
        }
      },
      "/usuarios/{usuarioId}": {
        "get": {
          "tags": [
            "usuario"
          ],
          "summary": "Busca um usuário pelo ID",
          "description": "Busca um único usuário",
          "operationId": "buscarUsuarioPorId",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a ser retornado",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Operação bem sucedida",
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            },
            "400": {
              "description": "Id inválido fornecido"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        },
        "put": {
          "tags": [
            "usuario"
          ],
          "summary": "Atualiza um usuário no LAS-API com os dados do formulário",
          "operationId": "atualizarUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a ser atualizado",
              "required": true,
              "type": "integer"
            },
            {
              "in": "body",
              "name": "usuario",
              "description": "Usuário a ser atualizado",
              "required": false,
              "schema": {
                "$ref": "#/definitions/Usuario"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Usuário atualizado com sucesso"
            },
            "405": {
              "description": "Entrada inválida"
            }
          }
        },
        "delete": {
          "tags": [
            "usuario"
          ],
          "summary": "Exclui um usuário",
          "operationId": "excluirUsuario",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a ser excluído",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "204": {
              "description": "Usuário excluído com sucesso"
            },
            "400": {
              "description": "Id inválido fornecido"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        }
      },
      "/usuarios/{usuarioId}/dados-pessoais": {
        "get": {
          "tags": [
            "dados-pessoais"
          ],
          "summary": "Busca os dados pessoais do usuário",
          "operationId": "buscarDadosPessoaisDoUsuario",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a retornar os dados pessoais",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Operação bem sucedida",
              "schema": {
                "$ref": "#/definitions/DadosPessoais"
              }
            },
            "400": {
              "description": "Id inválido fornecido"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        },
        "put": {
          "tags": [
            "dados-pessoais"
          ],
          "summary": "Atualiza os dados pessoais do usuário",
          "operationId": "atualizarDadosPessoaisDoUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a ser atualizado",
              "required": true,
              "type": "integer"
            },
            {
              "in": "body",
              "name": "dadosPessoais",
              "description": "Dados pessoais do usuário a ser atualizado",
              "required": false,
              "schema": {
                "$ref": "#/definitions/DadosPessoais"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Dados pessoais atualizados com sucesso"
            },
            "405": {
              "description": "Entrada inválida"
            }
          }
        }
      },
      "/usuarios/{usuarioId}/contatos": {
        "get": {
          "tags": [
            "contatos"
          ],
          "summary": "Busca os contatos do usuário",
          "operationId": "buscarContatosDoUsuario",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a retornar os contatos",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Operação bem sucedida",
              "schema": {
                "$ref": "#/definitions/Contatos"
              }
            },
            "400": {
              "description": "Id inválido fornecido"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        },
        "put": {
          "tags": [
            "contatos"
          ],
          "summary": "Atualiza os contatos do usuário",
          "operationId": "atualizarContatosDoUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a ser atualizado",
              "required": true,
              "type": "integer"
            },
            {
              "in": "body",
              "name": "contatos",
              "description": "Contatos do usuário a ser atualizado",
              "required": false,
              "schema": {
                "$ref": "#/definitions/Contatos"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Contatos atualizados com sucesso"
            },
            "405": {
              "description": "Entrada inválida"
            }
          }
        }
      },
      "/usuarios/{usuarioId}/endereco": {
        "get": {
          "tags": [
            "endereco"
          ],
          "summary": "Busca o endereço do usuário",
          "operationId": "buscarEndereçoDoUsuario",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a retornar o endereço",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Operação bem sucedida",
              "schema": {
                "$ref": "#/definitions/Endereco"
              }
            },
            "400": {
              "description": "Id inválido fornecido"
            },
            "404": {
              "description": "Usuário não encontrado"
            }
          }
        },
        "put": {
          "tags": [
            "endereco"
          ],
          "summary": "Atualiza o endereço do usuário",
          "operationId": "atualizarEnderecoDoUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a ser atualizado",
              "required": true,
              "type": "integer"
            },
            {
              "in": "body",
              "name": "contatos",
              "description": "Endereço do usuário a ser atualizado",
              "required": false,
              "schema": {
                "$ref": "#/definitions/Endereco"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Endereço atualizado com sucesso"
            },
            "405": {
              "description": "Entrada inválida"
            }
          }
        }
      },
...
      "/usuarios/{usuarioId}/senha": {
        "put": {
          "tags": [
            "senha"
          ],
          "summary": "Atualiza a senha do usuário",
          "operationId": "atualizarSenhaDoUsuario",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "usuarioId",
              "in": "path",
              "description": "Id do usuário a ser atualizado",
              "required": true,
              "type": "integer"
            },
            {
              "in": "body",
              "name": "senha",
              "description": "Senha do usuário a ser atualizado",
              "required": false,
              "schema": {
                "$ref": "#/definitions/Senha"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Senha atualizada com sucesso"
            },
            "405": {
              "description": "Entrada inválida"
            }
          }
        }
      },
...
    },
    "definitions": {
      "Usuario": {
        "type": "object",
        "required": [
          "id",
          "nome",
          "urlFotoPerfil"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "example": 12345
          },
          "nome": {
            "type": "string",
            "example": "Paulo Silva",
            "minLength": 5
          },
          "urlFotoPerfil": {
            "type": "string",
            "format": "uri",
            "example": "https://randomuser.me/api/portraits/men/91.jpg"
          }
        }
      },
...
      "DadosPessoais": {
        "type": "object",
        "required": [
          "cpf",
          "dataNascimento",
          "nomeCompleto",
          "rg"
        ],
        "properties": {
          "nomeCompleto": {
            "type": "string",
            "example": "Fulano de Tal",
            "minLength": 5
          },
          "dataNascimento": {
            "type": "string",
            "format": "date",
            "example": "2017-07-21T00:00:00.000+0000"
          },
          "rg": {
            "type": "string",
            "example": "123321 SSP BA",
            "minLength": 5
          },
          "cpf": {
            "type": "string",
            "example": "12312312312"
          }
        }
      },
      "Endereco": {
        "type": "object",
        "required": [
          "bairro",
          "cep",
          "endereco",
          "numero"
        ],
        "properties": {
          "cep": {
            "type": "string",
            "example": "72980000",
            "minLength": 8
          },
          "endereco": {
            "type": "string",
            "example": "Rua 123"
          },
          "numero": {
            "type": "integer",
            "example": 23
          },
          "complemento": {
            "type": "string",
            "example": "Apartamento 509"
          },
          "bairro": {
            "type": "string",
            "example": "Zona Norte"
          }
        }
      },
      "Contatos": {
        "type": "object",
        "required": [
          "celular",
          "email",
          "telefone"
        ],
        "properties": {
          "telefone": {
            "type": "string",
            "example": "6233311212"
          },
          "celular": {
            "type": "string",
            "example": "62998757575"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "foobar@gmail.com",
            "minLength": 5
          }
        }
      },
      "Senha": {
        "type": "object",
        "required": [
          "senha"
        ],
        "properties": {
          "senha": {
            "type": "string",
            "example": "admin",
            "minLength": 5
          }
        }
      },
...
    }
  }